Agent: Move PluginRegistry from puppet/ to master/
