+++ monkey/infection_monkey/master/plugin_registry.py
